ml-service/requirements-real.txt
# Child Growth Monitor ML Service - Real Production Requirements
# These are working, tested versions for actual anthropometric analysis

# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# Essential ML and computer vision libraries
opencv-python==4.8.1.78
mediapipe==0.10.8
Pillow==10.1.0
numpy==1.24.4

# Basic ML for anthropometric calculations
scikit-learn==1.3.2
joblib==1.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Optional: Advanced ML libraries (comment out if not needed)
# tensorflow==2.15.0  # For advanced models
# torch==2.1.2        # Alternative to TensorFlow
# scipy==1.11.4       # Scientific computing
